logic exception handling

+++ A3C6TV_HFT_2023241.Logic/Classes/BookingLogic.cs
@@ -1,5 +1,6 @@
 ﻿using A3C6TV_HFT_2023241.Models;
 using A3C6TV_HFT_2023241.Repository;
+using System;
 using System.Linq;
 
 namespace A3C6TV_HFT_2023241.Logic
@@ -18,18 +19,31 @@ namespace A3C6TV_HFT_2023241.Logic
         }
         public void Delete(int id)
         {
+            if (repo.Read(id) == null)
+            {
+                throw new ArgumentException("Booking with this ID doesn't exist!\nCannot delete it!");
+            }
             repo.Delete(id);
         }
         public Booking Read(int id)
         {
+            if (repo.Read(id) == null)
+            {
+                throw new ArgumentException("Customer with this ID doesn't exist!\nCannot read it!");
+            }
             return repo.Read(id);
         }
         public IQueryable<Booking> ReadAll()
         {
+            //kell a 0 eset?
             return repo.ReadAll();
         }
         public void Update(Booking item)
         {
+            if (repo.Read(item.BookingId) == null)
+            {
+                throw new ArgumentException("Booking with this ID doesn't exist!\nCannot update it!");
+            }
             repo.Update(item);
         }
